Module Diff.Make

Functor that produces a Diff module given a comparable type

Parameters

X : Comparable

Signature

type item = X.t

type diff =

| Deleted of item list

| Added of item list

| Equal of item list

type t = diff list

val get_diff : item array -> item array -> t

This is the main function : get_diff a1 a2 compares two arrays of items and outputs a list of chunks tagged with Deteted, Added or Removed