Module Legifrance_catala.Main

type new_article_version =

| NotAvailable

| Available of string

val check_article_expiration : Catala.Ast.law_article -> Api.access_token -> new_article_version option

Returns the ID of the future version of the article if any

type article_text_acc = {

article_title : string Catala.Pos.marked;

text : string;

new_version : string option;

current_version : string option;

}

Accumulator type when traversing the Catala source files

module Diff : sig ... end

Diff algorithm for a list of words

val compare_article_to_version : Api.access_token -> string -> string -> Diff.t option

compare_article_to_version token text version retrieves the text of the article whose LegiFrance ID is version and produces a diff with the expected text

val compare_to_versions : article_text_acc -> Api.access_token -> unit

Compares article_text_acc.current_version and article_text_acc.new_version by accessing LegiFrance and display differences if any

val include_legislative_text : string Catala.Pos.marked -> Api.access_token -> unit

Fill an @@Include ...@@ tag inside the Catala source file with the legislative contents retrieved from LegiFrance

val driver : string -> bool -> string -> string -> 'a

Parses the Catala master source file and checks each article:

if the article has a LegiFrance ID, checks the text of the article in the source code vs the text from LegiFrance;

if the article has an expiration date, display the difference between the current version of the article and the next one on LegiFrance;

fill each @@Include ...@@ tag with the contents retrieved from LegiFrance

val main : unit -> unit

Hook for the executable